prepare ingester for multiple integration tests
Summary:
We need to make sure that the init.sql is run for every integration environment.
The way we decided to do this is by running it as part of the beginning of the
integration test. It has to be idempotent so there won't be errors when running
it on existing tables. It has to be in the ami so that the integration test
can get the version of the .sql file that corresponds to the ingester code that's
being tested.

Test Plan:
ami built with no errors, and running the sql file against integration db caused
no errors

Reviewers: rhu

Reviewed By: rhu

Subscribers: jackgao, dwe, gareth, maasted

Differential Revision: https://phabricator.internal.justin.tv/D3290