ingester: log -> logrus logger - add rollbar logger
Summary:
Two logical changes in this, but the latter is small so grouping them together:
1. Switch to logrus logging
2. Use the logrus logger with rollbar hook and guarded goroutines

Adding rollbar to the logger here: https://phabricator.internal.justin.tv/D3844

Before landing I need to update aws_utils in _vendor in this repo. Didn't
include that yet since it wouldn't be pointing at aws_utils master yet since
that diff was landed, and it would add noise to this diff.

Test Plan:
Ran it locally and inserted panics and log errors in main and in goroutines and
saw the logs in stdout and in rollbar. Will integration test once the aws_utils
branch lands.

Reviewers: clgroft, dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D3860