Rollback transaction on error.
Summary:
Transaction wasn't being closed on error. Causes https://twitchtv.atlassian.net/browse/SCIENG-430
because of error when table doesn't exist.
Line279 was returning without closing the transaction, and it was retried 10
times. Each retry would completely lock a connection in the pool, so when the
pool was fully locked, the next `isolateTransaction` would hang, presumably
waiting for a member of the pool to be ready. This caused SIGINTs to not have
a place to actually exit the process, so it wouldn't restart without a hard
kill.

New behavior on missing table is to retry forever, erroring each time, and then
actually exit on SIGINT. It also starts behaving normally without restart if the
table is created during the endless retries.

Test Plan:
Verified the behavior I described above locally and on integration.
Integration test passes.

Reviewers: maasted

Reviewed By: maasted

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4303