ingester: fix bug transaction bug
Summary:
change staleLoadMetadata to not commit/rollback the transaction, have the calling function handle that
also changed a log msg to be less confusing in the most common case.

Test Plan:
ran integration test, passed
gave the int env an SQS message that would fail, it failed, but the db didn't
get locked and the thread didn't hang despite checking up on it and retrying it.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4366