Adding the transformer userIDWithMapping as bigint
Summary: In preparation for name changes being made into a public feature, this commit brings the new transformer mapping of userIDWithMapping as bigint. This transformer sets the value of a column expected to be a user id when is empty/invalid, and for that it expects a supporting column with a login so we can use it to map to an id through an underlying API.

Test Plan: Ran with a modified integration that sends events that use the new type and verified it ingested. For production, I intend to modify a single user_id in a low risk table to force the use of the transformer and verify it keeps ingesting correctly.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4779