add table name etc to log messages on table loads/errors
Summary:
When I saw the rollbar log message today for a failed load, I wanted to know
what table it was without querying ingesterdb/ingester logs for the
manifestuuid. The info is there at log time, so adding it to the log here for
the future.

Test Plan:
didn't test it on integration yet because can't build rn, but it's very simple
so I'm confident it's right. Going to test before I deploy to prod for sure.

Reviewers: clgroft

Reviewed By: clgroft

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5042