Tracking stats for all types of pending loads
Summary: Refactored the code to track same group metrics for in queue, stale events. In the process added the definition of tables pending migration, which are events in queue with a version larger than what’s currently recorded.

Test Plan: Ran on integration and updated a personal dashboard to confirm values. On prod, I'll use the same dashboard to confirm in combination with queries in ingester db.

Reviewers: clgroft

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5098