ingester: check if a failed load was commited before retrying
Summary: On retrieving failed loads to retry, test that they didn't actually get commited first, and mark them done if they were.

Test Plan:
I don't have a good way of testing that a redshift crash during a commit has the
behavior that the commit goes through but an error returns in the client on the
commit. However, assuming that is the case, I can test that this new logic will
mark the load as a success instead of retrying it.

I did this by:
- run integration test, watch logs
- find uuid of a successful load
- regenerate the tsv and manifest rows for that load so it looks like it was
  a failed load
- watch the rollbar error appear saying it ws marked as success
- see the rows disappear in manifest and tsv

This manual test passed. The integration test also passed.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5736