Create views in logs schema and write to physical schema
Summary:
Instead of writing to tables in the `logs` schema, write to
tables in the `physical` schema, dropping and recreating views
in `logs` (views can't be replaced if the number of columns
changes).
Have a view filter that applies to all tables with a
certain column.

Test Plan:
Ran integration test.

Deploy Plan:
1. Create the `mirror` schema (for jobs etl tables).
2. Stop ingester.
3. Deploy new config to ingester with view schema.
4. Disable all jobs that write to the `logs` schema.
5. Rename the `logs` schema to `physical`.
6. Start ingester.
7. Run a script (to be written) to create all current views
   in the logs schema and move all jobs logs tables to
   `db_snapshots` schema.
8. Deploy changes to jobs and restart disabled jobs.
9. Run enforce_perms script.

Reviewers: jackgao

Reviewed By: jackgao

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5824