Fix losing manifests on Redshift errors
Summary:
If an error occurs determining if a manifest is retryable,
the manifest will become ignored, with a null retry_ts.
Fix that by moving the fetch and update in with the retryable check.

There are still later errors that could occur, so I've ticketed
a better fix, but this catches this case of Redshift being down
for an extended period.

I also cleaned up some error messages.

Test Plan:
Ran integration test and tested loading with
an unavailable Redshift.

Reviewers: jackgao

Reviewed By: jackgao

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5837