Publish per-event last load time on ingester
Summary:
Ingester should publish when was the last time an event has been loaded so other systems that depend on freshness of data can get this information through HTTP request.

We introduce a new table in ingesterdb `last_load` that tracks the last_load of each event. Upon ingester startup, it pulls that table into an in-memory map. On load success, we update both the in memory structure as well as the database table for persistence.

Test Plan: `go test ./...`,  Tested on an instrumented variant of integration test.

Reviewers: ojcm

Reviewed By: ojcm

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5971